fix: examples also in ./schemas
